color backtrace works again!
The introduction of frame decorators had broken things, but if we bypass their
helpful nature, all is well.

The good news/bad news is that we are bypassing any helpful frame decorators.
The only one I know of right now is tromey's colorizing decorator at
https://github.com/tromey/gdb-helpers/blob/master/gdbhelpers/colorize.py
that is cool but suggests frame decorators are too limiting for our needs.  As
it happens, by bypassing them, we avoid interacting with the colorizer at all.